// libs/features/contributor/src/lib/shared/enums/form-wizard-types.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

/* Form wizard types enum */

export enum FormWizardTypes {
  PERSON_DETAILS = 'CONTRIBUTOR.SEC-PERSONAL',
  ENGAGEMENT_DETAILS = 'CONTRIBUTOR.SEC-ENGAGEMENT',
  DOCUMENT_DETAILS = 'CONTRIBUTOR.SEC-DOCUMENTS',
  CONTRACT_INFORMATIONS = 'CONTRIBUTOR.SEC-INFORMATIONS',
  CONTRACT_CLAUSES = 'CONTRIBUTOR.SEC-CLAUSES',
  PREVIEW = 'CONTRIBUTOR.PREVIEW',
  HEALTH_RECORD_DETAILS = 'CONTRIBUTOR.ADD-VIC.SEC-HEALTH-RECORD',
  TERMINATE_VIC_DETAILS = 'CONTRIBUTOR.TERMINATE-VIC.TERMINATE-VIC-DETAILS',
  CANCEL_VIC_DETAILS = 'CONTRIBUTOR.CANCEL-VIC.CANCEL-VIC-DETAILS',
  VERIFY_AUTHORIZATION = 'CONTRIBUTOR.VERIFY',
  AUTHORIZATION_DETAILS = 'CONTRIBUTOR.ADD-AUTHORIZATION.DETAILS',
  REQUEST_DETAILS = 'CONTRIBUTOR.SEC-REQUEST',
  VERIFY = 'CONTRIBUTOR.VERIFY'
}