// libs/features/contributor/src/lib/shared/enums/transaction-id.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum TransactionId {
  REGISTER_CONTRIBUTOR = 100609,
  CHANGE_ENGAGEMENT = 300330,
  MANAGE_WAGE = 300303,
  TERMINATE_CONTRIBUTOR = 250525,
  REGISTER_SECONDMENT = 101588,
  REGISTER_STUDY_LEAVE = 101589,
  CANCEL_CONTRIBUTOR = 100642,
  ADD_SECONDED = 100681,
  TRANSFER_CONTRIBUTOR = 100626,
  TRANSFER_ALL_CONTRIBUTOR = 100627,
  REGISTER_VIC_CONTRIBUTOR = 100624,
  MANAGE_VIC_WAGE = 100645,
  TERMINATE_VIC = 300333,
  CANCEL_VIC = 300334,
  ADD_CONTRACT = 300023,
  CANCEL_CONTRACT = 101003,
  CANCEL_ENGAGEMENT_VIOLATION = 101000,
  CHANGE_ENGAGEMENT_VIOLATION = 101001,
  TERMINATE_ENGAGEMENT_VIOLATION = 101002,
  ADD_AUTHORIZATION = 300355,
  MAINTAIN_COVERAGE = 100631,
  E_REGISTER_ENGAGEMENT = 101004,
  REACTIVATE_ENGAGEMENT = 101005,
  REACTIVATE_VIC_ENGAGEMENT = 101006,
  ENTER_RPA_AGGREGATION_FIRST_SCHEMA = 300001,
  ENTER_RPA_AGGREGATION_LAST_SCHEMA = 300002
}