// libs/features/contributor/src/lib/shared/enums/transaction-name.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum TransactionName {
  REGISTER_CONTRIBUTOR = 'Register Contributor',
  MANAGE_WAGE = 'Manage Wage',
  CHANGE_ENGAGEMENT = 'Change Engagement',
  TERMINATE_CONTRIBUTOR = 'Terminate Engagement',
  SECONDMENT_ENGAGEMENT = 'Register Secondment',
  STUDY_LEAVE_ENGAGEMENT = 'Register Study Request',
  CANCEL_CONTRIBUTOR = 'Cancel Engagement',
  TRANSFER_CONTRIBUTOR = 'Transfer Engagement',
  MANAGE_VIC_WAGE = 'Manage VIC Wage',
  TERMINATE_VIC = 'Terminate VIC',
  CANCEL_VIC = 'Cancel VIC',
  REGISTER_VIC = 'Register VIC Contributor',
  MAINTAIN_COVERAGE = 'Maintain Coverage'
}