// libs/features/contributor/src/lib/shared/models/vic-wage-update-summary.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

import { VicEngagementPeriod } from './vic-engagement-period';

export class VicWageUpdateSummary {
  currentPeriod: VicEngagementPeriod = new VicEngagementPeriod();
  updatedPeriod: VicEngagementPeriod = new VicEngagementPeriod();
}